Renaming module 'millis' to 'MillisFake' to better reflect the actual nature of the module
